Move graph utils
